Corrected half bug.
Corrected part of a bug that was allowing movement player when an event
was active (problem with event pages in the map "insidefarm"). There is
still half of the bug visible in the map "route", where you can move
while the soldier talks to you. (it's funny, because you can't move when
he is moving).